Fixes to MSVC warnings.  The original code made unwarranted assumptions about type equivalence: ints equaling shorts or chtypes and such.  This was probably not causing bugs,  but it _was_ quite definitely wrong and could have led to grief somewhere.

+++ demos/firework.c
@@ -34,7 +34,7 @@ int main(int argc, char **argv)
         start_color();
 
     for (i = 0; i < 8; i++)
-        init_pair(i, color_table[i], COLOR_BLACK);
+        init_pair( (short)i, color_table[i], COLOR_BLACK);
 
     seed = time((time_t *)0);
     srand(seed);
